Revert "Try maked a DCR interface"
This reverts commit 347273e719fa65d90c2692eed36e438e760e4cb6.
